refactor: 주문/배송 서비스 구조를 도메인 중심으로 통합
