--- order-service/src/main/java/com/palja/order_service/application/service/OrderDeliveryManagerService.java
@@ -1,9 +1,0 @@
-package com.palja.order_service.application.service;
-
-import com.palja.order_service.application.command.UpdateDeliveryStatusCommand;
-import com.palja.order_service.application.dto.response.DeliveryStatusRes;
-
-public interface OrderDeliveryManagerService {
-
-    DeliveryStatusRes updateDeliveryStatus(UpdateDeliveryStatusCommand command);
-}
