--- order-service/src/main/java/com/palja/order_service/application/service/OrderManagerService.java
@@ -1,8 +1,0 @@
-package com.palja.order_service.application.service;
-
-import com.palja.order_service.application.command.OrderStatusChangeCommand;
-import com.palja.order_service.application.dto.response.OrderStatusChangeRes;
-
-public interface OrderManagerService {
-    OrderStatusChangeRes changeOrderStatus(OrderStatusChangeCommand command);
-}
