+++ order-service/src/main/java/com/palja/order_service/presentation/controller/impl/OrderManagerControllerImpl.java
@@ -5,7 +5,7 @@ import com.palja.common.auditor.CurrentUser;
 import com.palja.common.response.ApiResponse;
 import com.palja.common.vo.UserRole;
 import com.palja.order_service.application.dto.response.OrderStatusChangeRes;
-import com.palja.order_service.application.service.OrderManagerService;
+import com.palja.order_service.application.service.OrderService;
 import com.palja.order_service.presentation.controller.OrderManagerController;
 import com.palja.order_service.presentation.dto.request.OrderStatusChangeReq;
 import jakarta.validation.Valid;
@@ -20,7 +20,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class OrderManagerControllerImpl implements OrderManagerController {
 
-    private final OrderManagerService orderManagerService;
+    private final OrderService orderService;
 
     // 주문 상태 변경 (관리자)
     @PutMapping("/{orderId}/status")
@@ -29,7 +29,7 @@ public class OrderManagerControllerImpl implements OrderManagerController {
             @PathVariable UUID orderId,
             @Valid @RequestBody OrderStatusChangeReq request
     ) {
-        OrderStatusChangeRes response = orderManagerService.changeOrderStatus(request.toCommand(orderId, CurrentUser.getLoginId()));
+        OrderStatusChangeRes response = orderService.changeOrderStatus(request.toCommand(orderId, CurrentUser.getLoginId()));
         return ResponseEntity.ok(ApiResponse.success(response, "주문 상태가 변경되었습니다."));
     }
 }
